implemented entry deletion, added real foreign_key logic

+++ lib/data/repositories/category_repository.dart
@@ -1,4 +1,5 @@
 import 'package:drift/drift.dart';
+import 'package:organizer/core/utils/json_util.dart';
 import 'package:organizer/data/database/database.dart' hide Category;
 import 'package:organizer/core/models/category.dart';
 import 'package:organizer/core/repositories/category_repository.dart';
@@ -31,8 +32,32 @@ class CategoryRepositoryDrift implements CategoryRepository {
 
   @override
   Future<void> removeCategory(int categoryId) async {
-    await (db.delete(
-      db.categories,
-    )..where((c) => c.id.equals(categoryId))).go();
+    await db.transaction(() async {
+      final topics = await (db.select(
+        db.topics,
+      )..where((t) => t.categoryId.equals(categoryId))).get();
+      final topicIds = topics.map((t) => t.id).toList();
+      
+      final varTransactions = await (db.select(
+        db.varTransactions,
+      )..where((v) => v.topicId.isIn(topicIds))).get();
+
+      for (final v in varTransactions) {
+        if (v.compensations != null) {
+          final compensations = JsonUtil.string2CompensationInfo(
+            v.compensations,
+          )!;
+          for (final cId in compensations.keys) {
+            await (db.delete(
+              db.varTransactions,
+            )..where((vT) => vT.id.equals(cId))).go();
+          }
+        }
+      }
+
+      await (db.delete(
+        db.categories,
+      )..where((c) => c.id.equals(categoryId))).go();
+    });
   }
 }
